fix nt passing unit tests
