+++ libs/infinity_emb/infinity_emb/transformer/utils_optimum.py
@@ -1,6 +1,8 @@
 from pathlib import Path
 from typing import List, Union
 
+import numpy as np
+
 from infinity_emb.log_handler import logger
 
 try:
@@ -17,6 +19,14 @@ try:
     import torch
 except ImportError:
     torch = None  # type: ignore
+
+
+def normalize(input_array, p=2, dim=1, eps=1e-12):
+    # Calculate the Lp norm along the specified dimension
+    norm = np.linalg.norm(input_array, ord=p, axis=dim, keepdims=True)
+    norm = np.maximum(norm, eps)  # Avoid division by zero
+    normalized_array = input_array / norm
+    return normalized_array
 
 
 def device_to_onnx(device) -> Union[str, List[str]]:
