+++ libs/infinity_emb/tests/unit_test/inference/test_models.py
@@ -1,6 +1,7 @@
 """
 Tests that the pretrained models produce the correct scores on the STSbenchmark dataset
 """
+
 import copy
 from typing import List
