+++ libs/infinity_emb/tests/unit_test/inference/test_select_model.py
@@ -9,9 +9,11 @@ from infinity_emb.transformer.utils import InferenceEngine
 def test_engine(engine):
     select_model(
         engine=engine,
-        model_name_or_path="TaylorAI/bge-micro-v2"
-        if engine == InferenceEngine.optimum
-        else pytest.DEFAULT_BERT_MODEL,
+        model_name_or_path=(
+            "TaylorAI/bge-micro-v2"
+            if engine == InferenceEngine.optimum
+            else pytest.DEFAULT_BERT_MODEL
+        ),
         batch_size=4,
         device=Device.cpu,
     )
